requirements.txt
tensorflow>=2.8
opencv-python
ffmpeg-python
numpy
pillow
tqdm
python-dotenv
inference
ultralytics
torch>=1.10.0
torchvision>=0.11.0
scipy
# DeepSort tracking
deep-sort-realtime
google-genai